requirements.txt
streamlit
requests
timezonefinder
pytz
Pillow
plotly
pandas
numpy
matplotlib
geosky
countryinfo
Flask